HEALTHAI/Project Files/requirements.txt
streamlit
pandas
numpy
plotly
requests
python-dotenv
ibm-watson-machine-learning
ibm-watsonx-ai